fixed bigint fk in schema.sql

+++ src/main/resources/schema.sql
@@ -21,7 +21,7 @@ CREATE UNIQUE INDEX users_unique_email_idx ON users (email);
 
 CREATE TABLE user_roles
 (
-    user_id INTEGER NOT NULL,
+    user_id bigint NOT NULL,
     role    VARCHAR(255),
     CONSTRAINT user_roles_idx UNIQUE (user_id, role),
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
@@ -40,7 +40,7 @@ CREATE TABLE dishes
     name        VARCHAR(255) NOT NULL,
     price       INTEGER NOT NULL,
     day   DATE DEFAULT now() NOT NULL,
-    restaurant_id INTEGER NOT NULL,
+    restaurant_id bigint NOT NULL,
     CONSTRAINT name_day_restaurant_idx UNIQUE (name, day, restaurant_id),
     FOREIGN KEY (restaurant_id) REFERENCES restaurants (id) ON DELETE CASCADE
 );
@@ -50,8 +50,8 @@ CREATE TABLE votes
     id   bigint default global_seq.nextval primary key,
     day  DATE DEFAULT now() NOT NULL,
     time TIME DEFAULT now() NOT NULL,
-    user_id    INTEGER NOT NULL,
-    restaurant_id    INTEGER NOT NULL,
+    user_id    bigint NOT NULL,
+    restaurant_id    bigint NOT NULL,
     CONSTRAINT vote_per_day_idx UNIQUE (day, user_id),
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE,
     FOREIGN KEY (restaurant_id) REFERENCES restaurants (id) ON DELETE RESTRICT
